Punto intermedio .__.

+++ src/components/Componentes_usuario/ListaUsuarios.vue
@@ -4,10 +4,12 @@
       <div class="card-header"><strong>Lista de Usuarios</strong></div>
 
       <div class="card_body">
+        <button type="button" v-on:click="  editaregistro()" class="btn btn-outline-warning">Editar</button>
         <table class="table">
+          
           <thead>
             <tr>
-              <th style="font-weight: bold;display: none;">Id</th>
+              <th style="font-weight: bold;">Id</th>
               <th style="font-weight: bold;">user</th>
               <th style="font-weight: bold;">password</th>
               <th style="font-weight: bold;">fechaRegistro</th>
@@ -18,7 +20,7 @@
           </thead>
           <tbody>
             <tr v-for="usuario in Usuarios" :key="usuario.pkUsuario">
-              <td style="display: none;">{{ usuario.pkUsuario }}</td>
+              <td>{{ usuario.pkUsuario }}</td>
               <td>{{ usuario.user }}</td>
               <td>{{ usuario.password }}</td>
               <td>{{ usuario.fechaRegistro }}</td>
@@ -28,14 +30,14 @@
                 <div class="btn-group" role="label" aria-label="">
                   <!-- |<router-link :to="{name:'editar',param:{id:articulo.id}}" class="btn btn-info">Editar</router-link> | -->
                   <button type="button" v-on:click="borrarArticulo(usuario.pkUsuario)" class="btn btn-outline-danger">
-                    Eliminar</button>
+                    Eliminar</button>          
                 </div>
               </td>
+              
             </tr>
           </tbody>
         </table>
 
-        <div>{{ smg }}</div>
       </div>
 
 
@@ -45,23 +47,30 @@
 
 <script>
 import axios from "axios";
+
+
 export default {
+
+  components:{
+
+  },
   data() {
     return {
       Usuarios: [],
       Empleado: [],
       nombresempleados: [{}],
-      smg: ""
+      smg: "",
     };
   },
   created: function () {
-    this.consultarArticulos();
+    this.consultarUsuario();
   },
   methods: {
-    consultarArticulos() {
+    consultarUsuario() {
       axios.get("https://localhost:7051/Usuario").then((result) => {
         console.log(result.data);
         this.Usuarios = result.data.result;
+
         // console.log(this.Usuarios[1].fkEmpleado);
 
 
@@ -89,20 +98,24 @@ export default {
     },
 
     borrarArticulo(id) {
-      console.log(id);
+      var pregunta=window.confirm('Esta se seguro de eliminar este registro?');
 
-      axios.delete("https://localhost:7051/Usuario/" + id);
-      if (id != 0) {
-        window.location.href = "dashboard";
-      } else {
-        this.smg = "No tiene valor mayor a 0"
-      }
+      if(pregunta===true){
+        axios.delete("https://localhost:7051/Usuario/" + id);
+        window.location.href = "get";
+        
+      }else{
+        
+      }  
 
     },
-    BuscarEmpleado() {
-    console.log("fuera de servicio :p")
+    editaregistro() {
+    window.location.href="/putUsuario";
+
+    
 
     },
   },
+
 };
 </script>
